Add fallback to environment variable if .env doesn't exist
In `get_api_key()` function, I added a check to see if the `OPENAI_APIKEY` is present in environment variables, as a fallback in case the `.env` file is not found.

This was necessary because some users may not have the `.env` file in their home directory or have chosen to set the `OPENAI_APIKEY` environment variable directly.